Speed up build by not requiring yast2 (bsc#999203).
This is pushing the limits a bit, as this is possible only because
this package does not run any tests during the build.

On the other hand, it has been a very stable package, that is sitting on
the [critical rebuild path][cp] (pam BuildRequires yast2 which Requires
perl-bindings), so this savings should be worth it.

[cp]: https://github.com/mvidner/rpm-build-dependencies/blob/762b46108330fe2ca6382fc4776b5c204f6e7c00/yast_deps.yaml#L9