Add a CFA model to manage nsswitch.conf file
With the new layout to separate the distribution provided configuration
files from the changes the admin makes, the NSS configuration is shipped
in /usr/etc/nsswitch.conf instead of /etc/nsswitch.conf.

However, adapting YaST is not a matter of changing the path to the file
in the SCR agent, since files located below /usr/etc are not modifiable
by the admin. Even more, this configuration file does not allow to
override a single entry in /etc/nsswitch.conf. Instead, it should
include all the desired configuration because once /etc/nsswitch.conf
exists /usr/etc/nsswitch.conf will be completely ignored.

The CFA class added in this set of changes ensures not losing the vendor
configuration using it as the starting point when the local
configuration file does not exist yet.

To know more, see

* https://github.com/yast/yast-yast2/pull/1082
* https://bugzilla.suse.com/show_bug.cgi?id=1173119#c2
* https://en.opensuse.org/openSUSE:Packaging_UsrEtc#Variant_3
* https://travis-ci.com/github/yast/usr-etc-test/jobs/342507898#L222